Changes implemented from Stereogram-iPad: Split the model into three classes. One to manage the images on disk, one to manipulate UIImage objects and one to hold a cache of thumbnails. I've also allowed the thumbnail cache to receive low-memory notifications and to delete itself when they are triggered. Changed PhotoStore so it is not a singleton. Now it has a constructor that takes an NSError object, so the view-controller that owns it can display errors if it fails to initialise. I've moved the property-finding code so that it is now called from that constructor, instead of being a public method.

+++ Stereogram/PWAlertView.m
@@ -73,7 +73,7 @@ confirmButtonTitle:(NSString *)confirmTitle
 {
     NSString *buttonTitle = [alertview buttonTitleAtIndex:buttonIndex];
     PWAlertView_Action action = [buttonTitlesAndBlocks objectForKey:buttonTitle];
-    NSAssert(action, @"No action for button title [%@] at index %d", buttonTitle, buttonIndex);
+    NSAssert(action, @"No action for button title [%@] at index %ld", buttonTitle, (long)buttonIndex);
     if(action)
         action();
     [self.class removeAReference:self];
